Rename `payas_model::import` to `payas_model::spec`
