+++ payas-model/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod import;
 pub mod model;
+pub mod spec;
 
 pub use payas_sql::sql;
